+++ dwds/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 24.2.1
 
 - Update to be forward compatible with changes to `package:shelf_web_socket`.
+- Added support for some debugging APIs with the DDC library bundle format. - [#2537](https://github.com/dart-lang/webdev/issues/2537)
 - Expose a partial implementation of
   `FrontendServerDdcLibraryBundleStrategyProvider`.